refactor: add button for new acc creation
