+++ packages/core/src/App/Components/Layout/Header/dtrader-v2/account-group-wrapper-dtrader-v2.tsx
