+++ packages/core/src/sass/app/_common/components/account-switcher-dtrader-v2.scss
@@ -33,7 +33,8 @@
         }
     }
 
-    &__account {
+    &__account,
+    &__new-account {
         display: flex;
         align-items: center;
         justify-content: space-between;
